osce-backend/requirement.txt
djangorestframework==3.16.1
Django==5.2.6
drf-spectacular==0.28.0
django-cors-headers == 4.9.0
drf-writable-nested==0.7.2
gunicorn==23.0.0
requests==2.32.5
arrow==1.3.0
pillow==11.3.0
yarl==1.22.0
rules==3.5
djangorestframework-simplejwt==5.5.1
python-dotenv==1.2.1
langchain-openai==1.1.7
faster-whisper==1.2.1
uvicorn==0.40.0
safetensors==0.7.0
scipy==1.17.0
scikit-image==0.26.0
facexlib==0.3.0
yacs==0.1.8
librosa==0.11.0
ultralytics==8.4.8
pydub==0.25.1
tensorboardx==2.6.4
transformers==5.0.0
kornia==0.8.2
numpy==2.3.0
aiortc==1.14.0
channels==4.3.2
uvicorn[standard]
edge-tts==7.2.7